Restructure initialization to happen outside of main()
This means the configuration is now a global struct, that the functions
exposed are defined elsewhere, and that initialization now happens in
cljs_engine_init.
