+++ planck-c/cljs.c
@@ -6,6 +6,9 @@
 #include <JavaScriptCore/JavaScript.h>
 
 #include "bundle.h"
+#include "functions.h"
+#include "globals.h"
+#include "http.h"
 #include "io.h"
 #include "jsc_utils.h"
 #include "str.h"
@@ -241,4 +244,102 @@ char **get_completions(JSContextRef ctx, const char *buffer, int *num_completion
 
 	*num_completions = n;
 	return completions;
+}
+
+void register_global_function(JSContextRef ctx, char *name, JSObjectCallAsFunctionCallback handler) {
+	JSObjectRef global_obj = JSContextGetGlobalObject(ctx);
+
+	JSStringRef fn_name = JSStringCreateWithUTF8CString(name);
+	JSObjectRef fn_obj = JSObjectMakeFunctionWithCallback(ctx, fn_name, handler);
+
+	JSObjectSetProperty(ctx, global_obj, fn_name, fn_obj, kJSPropertyAttributeNone, NULL);
+}
+
+void cljs_engine_init(JSContextRef ctx) {
+	JSStringRef nameRef = JSStringCreateWithUTF8CString("planck");
+	JSGlobalContextSetName(ctx, nameRef);
+
+	evaluate_script(ctx, "var global = this;", "<init>");
+
+	register_global_function(ctx, "AMBLY_IMPORT_SCRIPT", function_import_script);
+	bootstrap(ctx, config.out_path);
+
+	register_global_function(ctx, "PLANCK_CONSOLE_LOG", function_console_log);
+	register_global_function(ctx, "PLANCK_CONSOLE_ERROR", function_console_error);
+
+	evaluate_script(ctx, "var console = {};"\
+			"console.log = PLANCK_CONSOLE_LOG;"\
+			"console.error = PLANCK_CONSOLE_ERROR;", "<init>");
+
+	// require app namespaces
+	evaluate_script(ctx, "goog.require('planck.repl');", "<init>");
+
+	// without this things won't work
+	evaluate_script(ctx, "var window = global;", "<init>");
+
+	register_global_function(ctx, "PLANCK_READ_FILE", function_read_file);
+	register_global_function(ctx, "PLANCK_LOAD", function_load);
+	register_global_function(ctx, "PLANCK_LOAD_DEPS_CLJS_FILES", function_load_deps_cljs_files);
+	register_global_function(ctx, "PLANCK_CACHE", function_cache);
+
+	register_global_function(ctx, "PLANCK_EVAL", function_eval);
+
+	register_global_function(ctx, "PLANCK_GET_TERM_SIZE", function_get_term_size);
+	register_global_function(ctx, "PLANCK_PRINT_FN", function_print_fn);
+	register_global_function(ctx, "PLANCK_PRINT_ERR_FN", function_print_err_fn);
+
+	register_global_function(ctx, "PLANCK_SET_EXIT_VALUE", function_set_exit_value);
+
+	register_global_function(ctx, "PLANCK_RAW_READ_STDIN", function_raw_read_stdin);
+	register_global_function(ctx, "PLANCK_RAW_WRITE_STDOUT", function_raw_write_stdout);
+	register_global_function(ctx, "PLANCK_RAW_FLUSH_STDOUT", function_raw_flush_stdout);
+	register_global_function(ctx, "PLANCK_RAW_WRITE_STDERR", function_raw_write_stderr);
+	register_global_function(ctx, "PLANCK_RAW_FLUSH_STDERR", function_raw_flush_stderr);
+
+	register_global_function(ctx, "PLANCK_REQUEST", function_http_request);
+
+	{
+		JSValueRef arguments[config.num_rest_args];
+		for (int i = 0; i < config.num_rest_args; i++) {
+			arguments[i] = c_string_to_value(ctx, config.rest_args[i]);
+		}
+		JSValueRef args_ref = JSObjectMakeArray(ctx, config.num_rest_args, arguments, NULL);
+
+		JSValueRef global_obj = JSContextGetGlobalObject(ctx);
+		JSStringRef prop = JSStringCreateWithUTF8CString("PLANCK_INITIAL_COMMAND_LINE_ARGS");
+		JSObjectSetProperty(ctx, JSValueToObject(ctx, global_obj, NULL), prop, args_ref, kJSPropertyAttributeNone, NULL);
+		JSStringRelease(prop);
+	}
+
+	evaluate_script(ctx, "cljs.core.set_print_fn_BANG_.call(null,PLANCK_PRINT_FN);", "<init>");
+	evaluate_script(ctx, "cljs.core.set_print_err_fn_BANG_.call(null,PLANCK_PRINT_ERR_FN);", "<init>");
+
+	char *elide_script = str_concat("cljs.core._STAR_assert_STAR_ = ", config.elide_asserts ? "false" : "true");
+	evaluate_script(ctx, elide_script, "<init>");
+	free(elide_script);
+
+	{
+		JSValueRef arguments[4];
+		arguments[0] = JSValueMakeBoolean(ctx, config.repl);
+		arguments[1] = JSValueMakeBoolean(ctx, config.verbose);
+		JSValueRef cache_path_ref = NULL;
+		if (config.cache_path != NULL) {
+			JSStringRef cache_path_str = JSStringCreateWithUTF8CString(config.cache_path);
+			cache_path_ref = JSValueMakeString(ctx, cache_path_str);
+		}
+		arguments[2] = cache_path_ref;
+		arguments[3] = JSValueMakeBoolean(ctx, config.static_fns);
+		JSValueRef ex = NULL;
+		JSObjectCallAsFunction(ctx, get_function(ctx, "planck.repl", "init"), JSContextGetGlobalObject(ctx), 4, arguments, &ex);
+		debug_print_value("planck.repl/init", ctx, ex);
+	}
+
+	if (config.repl) {
+		evaluate_source(ctx, "text", "(require '[planck.repl :refer-macros [apropos dir find-doc doc source pst]])", true, false, "cljs.user", "dumb");
+	}
+
+	evaluate_script(ctx, "goog.provide('cljs.user');", "<init>");
+	evaluate_script(ctx, "goog.require('cljs.core');", "<init>");
+
+	evaluate_script(ctx, "cljs.core._STAR_assert_STAR_ = true;", "<init>");
 }
