+++ planck-c/cljs.h
@@ -10,3 +10,5 @@ void run_main_in_ns(JSContextRef ctx, char *ns, int argc, char **argv);
 
 char *get_current_ns();
 char **get_completions(JSContextRef ctx, const char *buffer, int *num_completions);
+
+void cljs_engine_init(JSContextRef ctx);
