+++ planck-c/jsc_utils.h
@@ -1,6 +1,15 @@
 #include <JavaScriptCore/JavaScript.h>
 
 JSStringRef to_string(JSContextRef ctx, JSValueRef val);
+
+#ifdef DEBUG
+#define debug_print_value(prefix, ctx, val)	print_value(prefix ": ", ctx, val)
+#else
+#define debug_print_value(prefix, ctx, val)
+#endif
+
+void print_value(char *prefix, JSContextRef ctx, JSValueRef val);
+
 JSValueRef evaluate_script(JSContextRef ctx, char *script, char *source);
 
 char *value_to_c_string(JSContextRef ctx, JSValueRef val);
